Core

Fermi.GaussianBasis.BasisFunction — Type

Fermi.GaussianBasis.BasisFunction

Object representing a shell of Gaussian basis functions composed of $N$ primitives:

\[\chi_{l,m} = \sum_i^N C_i r^{l}e^{-\zeta_i r^2} Y_{l,m}(\theta,\phi)\]

Fields

Name	Type	Description

l	Int32	Angular momentum number (e.g. 0, 1, 2 for S, P, D...)

coef	Array{Float64,1}	Array with coefficients ($C_i$) for each primitive

exp	Array{Float64,1}	Array with exponents ($\zeta_i$) for each primitive

Examples

julia> bf = Fermi.GaussianBasis.BasisFunction(1, [1/√2, 1/√2], [5.0, 1.2])
P shell with 3 basis built from 2 primitive gaussians

χ₁₋₁ =    0.7071067812⋅Y₁₋₁⋅r¹⋅exp(-5.0⋅r²)
     +    0.7071067812⋅Y₁₋₁⋅r¹⋅exp(-1.2⋅r²)

χ₁₀  =    0.7071067812⋅Y₁₀⋅r¹⋅exp(-5.0⋅r²)
     +    0.7071067812⋅Y₁₀⋅r¹⋅exp(-1.2⋅r²)

χ₁₁  =    0.7071067812⋅Y₁₁⋅r¹⋅exp(-5.0⋅r²)
     +    0.7071067812⋅Y₁₁⋅r¹⋅exp(-1.2⋅r²)

Fermi.GaussianBasis.BasisSet — Type

Fermi.GaussianBasis.BasisSet

Object holding a set of BasisFunction objects associated with each atom in a molecule.

Fields

Name	Type	Description

molecule	Molecule	Fermi.Geometry.Molecule object

basis_name	String	String holding the basis set name

basis	Dict{Atom,Array{BasisFunction,1}}	A dictionary that maps Atom objects to an Array of BasisFunction

natoms	Int32	Number of atoms in the BasisSet

nbas	Int32	Number of basis functions (Note, not equal the number of BasisFunction objects)

nshells	Int32	Number of shells, i.e. BasisFunction objects

lc_atoms	Array{Int32,1}	Integer array maping data to libcint

lc_bas	Array{Int32,1}	Integer array maping data to libcint

lc_env	Array{Float64,1}	Float64 array maping data to libcint

Example

Build a basis set from default options

julia> bset = Fermi.GaussianBasis.BasisSet()
sto-3g Basis Set
Number of shells: 5
Number of basis:  7

O: 1s 2s 1p 
H: 1s 
H: 1s

The BasisSet object can be accessed as two-dimensional array.

julia> bset[2] # Show all basis for the second atom (H 1s)
1-element Vector{Fermi.GaussianBasis.BasisFunction}:
 S shell with 1 basis built from 3 primitive gaussians

χ₀₀  =    0.9817067283⋅Y₀₀⋅exp(-3.425250914⋅r²)
     +    0.9494640079⋅Y₀₀⋅exp(-0.6239137298⋅r²)
     +    0.2959064597⋅Y₀₀⋅exp(-0.168855404⋅r²)
julia> bset[1,2] # Show the second basis for the first atom (O 2s)
S shell with 1 basis built from 3 primitive gaussians

χ₀₀  =    0.8486970052⋅Y₀₀⋅exp(-5.033151319⋅r²)
     +    1.1352008076⋅Y₀₀⋅exp(-1.169596125⋅r²)
     +    0.8567529838⋅Y₀₀⋅exp(-0.38038896⋅r²)

You can also create your own crazy mix!

julia> @molecule {
    H 0.0 0.0 0.0
    H 0.0 0.0 0.7
} 
julia> mol = Fermi.Geometry.Molecule()
# Lets create an S basis function for H
julia> s = Fermi.GaussianBasis.BasisFunction(0, [0.5215367271], [0.122])
# Now a P basis function
julia> p = Fermi.GaussianBasis.BasisFunction(1, [1.9584045349], [0.727])
# Now to create the basis set we need a dictionary maping atoms do basis functions
# Let's fetch the atom list from the molecule object
julia> Atoms = mol.atoms
# Now create the desired mapping as a dictionary. We want the first hydrogen
# to use only the S function, whereas the second will use both S and P
julia> shells = Dict(
    Atoms[1] => [s],
    Atoms[2] => [s,p]
)
# Now we can create the basis set!
julia> Fermi.GaussianBasis.BasisSet(mol, "My Crazy", shells)
My Crazy Basis Set
Number of shells: 3
Number of basis:  5

H: 1s 
H: 1s 1p