Energy computations

Fermi.@energy — Macro

@energy(comm)

Macro to call functions to compute energy given current options.

Implemented methods:

RHF            Restricted Hartree-Fock
RMP2           Restricted Moller-Plesset PT order 2
RMP3           Restricted Moller-Plesset PT order 3
CCSD           Restricted Coupled-Cluster with Single and Double substitutions
CCSD(T)        Restricted Coupled-Cluster with Single and Double substitutions and perturbative triples.
ecCCSD         Restricted externally corrected CCSD. External correction from CASCI computation.
ecCCSD(T)      Restricted externally corrected CCSDpT. External correction from CASCI computation.