Integrals and Orbitals

IntegralHelper

The Fermi.Integrals.IntegralHelper object is designed to provide easy access to molecular integrals from Julia.

Creation

To create an IntegralHelper, you can simply call,

Fermi.Integrals.IntegralHelper()

and one will be created based on the information in Fermi.CurrentOptions, which you may recall are set with @set. If you want more control, you can pass keyword arguments for the Molecule and OrbDict to be used,

Fermi.Integrals.IntegralHelper(mol=mymol,orbs=myorbs)

Use

Once an IntegralHelper object is created , integrals are simply accessed as,

helper["S"] #AO basis overlap
helper["T"] #AO basis kinetic
helper["V"] #AO basis nuclear attraction
helper["\mu"] #AO basis ERI
helper["B"] #AO basis density fitted ERI (B(P,m,n))

helper["OOOO"] # MO basis <ij|kl> TEI's 
helper["OOVV"] # MO basis <ij|ab> TEI's
helper["(OVOV)"] # MO basis (ia|jb) TEI's **chemists notation**
helper["FOV"] # MO basis Fock matrix, OV block
helper["BOV"] # MO basis DF TEI's, (ia|P)

Direct access to the underlying orbitals is possible,

helper.orbs["O"]   # occupied orbitals, array of Orbital objects
helper.orbs["[O]"] # occupied orbitals, collected into a matrix

To obtain unit normalized (non-CCA) integrals for d and higher angular momentum functions, call Fermi.Integrals.normalize(h::IntegralHelper,true). Note that this will delete all current integrals.

Fermi.Integrals.IntegralHelper — Type

IntegralHelper{T}

Structure to assist with computing and storing integrals. Accesss like a dictionary e.g., ints["B"] to get DF ERI integrals. A number of characters can be added to denote orbitals in various bases, such as MO or NO. O -> occ,alpha,ERI o -> occ,beta,ERI V -> vir,alpha,ERI v -> vir,beta,ERI P -> all,alpha,ERI p -> all,beta,ERI B -> DF-ERI μ -> ao,ERI Ω -> NO,occ,alpha ω -> NO,occ,beta U -> NO,vir,alpha u -> NO,vir,beta S -> AO,overlap T -> AO,kinetic V -> AO,nuclear

Fields

cache::Dict{String,Array}                        previously computed integrals
bname::Dict{String,String}                       basis set names for various purposes
mol::Molecule                                    attached Molecule object.
orbs::Dict{String,O} where O < AbstractOrbitals  orbitals of various kinds
basis::Dict{String,Lints.BasisSetAllocated}      basis sets for various purposes

Fermi.Integrals.aokinetic — Function

aokinetic(molecule::Fermi.Molecule, basis::String)

Computes AO basis kinetic energy ⟨μ|T̂|ν⟩ integrals for the given basis and molecule. Can be accessed at a higher level by calling

helper["T"]

where helper is bound to the desired molecule and basis set.

Fermi.Integrals.aooverlap — Function

aooverlap(molecule::Fermi.Molecule, basis::String)

Computes AO basis overlap ⟨p|q⟩ integrals for the given basis and molecule. Can be accessed at a higher level by calling

helper["S"]

where helper is bound to the desired molecule and basis set.

Missing docstring.

Missing docstring for Fermi.Integrals.aodipole. Check Documenter's build log for details.

Fermi.Integrals.aonuclear — Function

aonuclear(molecule::Fermi.Molecule, basis::String)

Computes AO basis nuclear attraction ⟨μ|V̂|ν⟩ integrals for the given basis and molecule. Can be accessed at a higher level by calling

helper["V"]

where helper is bound to the desired molecule and basis set.

Fermi.Integrals.aoeri — Function

aoeri(molecule::Fermi.Molecule, basis::String)

Computes AO basis electron repulsion integrals ⟨μν|Ô₂|ρσ⟩ integrals for the given basis and molecule. Can be accessed at a higher level by calling

helper["μ"]

where helper is bound to the desired molecule and basis set.

Fermi.Integrals.dfaoeri — Function

dfaoeri(molecule::Fermi.Molecule, basis::String)

Computes AO basis density fitted electron repulsion integrals ⟨μν|Ô₂|P⟩J(P,Q)^-1/2 integrals for the given basis and molecule. Note that the returned integrals DO NOT need to be combined with the Coulomb metric J(P,Q). In common notation, this is B(Q,μ,ν). Can be accessed at a higher level by calling

helper["B"]

where helper is bound to the desired molecule and basis set.

Fermi.Integrals.aux_ri! — Function

aux_ri!(I::IntegralHelper, ri=Fermi.CurrentOptions["rifit"])

Clears the integral cache and switches auxiliary DF integrals to use the current RI fitting basis set. Used between DF-RHF and DF-post HF.

Fermi.Integrals.aux_jk! — Function

aux_ri!(I::IntegralHelper, jk=Fermi.CurrentOptions["jkfit"])

Clears the integral cache and switches auxiliary DF integrals to use the current JK fitting basis set. Used to ensure JK integrals are used in DF-RHF.

Base.getindex — Function

getindex(collection, key...)

Retrieve the value(s) stored at the given key or index within a collection. The syntax a[i,j,...] is converted by the compiler to getindex(a, i, j, ...).

Examples

julia> A = Dict("a" => 1, "b" => 2)
Dict{String,Int64} with 2 entries:
  "b" => 2
  "a" => 1

julia> getindex(A, "a")
1