Modules

Fermi has a number of submodules organized for specific tasks. All modules are accessed as Fermi.ModuleName.

Options

Fermi.Options — Module

Fermi.Options

Module to manage options in Fermi.

Functions

Fermi.Options.set(option, value)     Set an <option> to a given <value>
Fermi.Options.get(option)            Return the current value of an <option>
Fermi.Options.reset()                Reset all options to default values
Fermi.Options.reset(option)          Reset a specific <option> to its default value
Fermi.Options.molecule(molstring)    Read in a String for the `molstring` option

Alternatively, at global scope, one can use the corresponding macros that create shortcuts for the commands above

Macros

@set <option> <value>   Set an <option> to a given <value>
@get <option>           Return the current value of an <option>
@reset                  Reset all options to default values
@reset <option>         Reset a specific <option> to its default value
@molecule               Read in a String for the `molstring` option
@lookup <string>        Search for keywords containing <string>

Libcint

Fermi.Libcint — Module

Fermi.Libcint

Interface between Fermi and the integral library libcint. This module exposes libcint functions to the Julia interface.

DIIS

Missing docstring.

Missing docstring for Fermi.DIIS. Check Documenter's build log for details.

PhysicalConstants

Fermi.PhysicalConstants — Module

Fermi.PhysicalConstants

Module for storing physical constants and conversion factors used in computations.

Functions:

atomic_number   Given an element symbol, return the atomic number.

GaussianBasis

Fermi.GaussianBasis — Module

Fermi.GaussianBasis

Module handling basis functions and basis set.

Orbitals

Missing docstring.

Missing docstring for Fermi.Orbitals. Check Documenter's build log for details.

Integrals

Fermi.Integrals — Module

Fermi.Integrals

Module to compute and manage molecular integrals.

HartreeFock

Missing docstring.

Missing docstring for Fermi.HartreeFock. Check Documenter's build log for details.

MollerPlesset

Fermi.MollerPlesset — Module

Fermi.MøllerPlesset

Module for running Møller–Plesset perturbation theory computations.

CoupledCluster

Fermi.CoupledCluster — Module

Fermi.CoupledCluster

Module for running CoupledCluster computations in Fermi.